+++ src/components/Description.js
@@ -1,5 +1,8 @@
 import PropTypes from 'prop-types';
 import React from 'react';
+import Markdown from 'react-remarkable';
+
+const mdOptions = { breaks: true };
 
 // parse description string to detect where to add links on tracker
 const getTrackerLinks = (description, pattern = '#([0-9]+)', trackerUrl) => {
@@ -12,28 +15,22 @@ const getTrackerLinks = (description, pattern = '#([0-9]+)', trackerUrl) => {
   let index = 0;
   while ((match = reIssueId.exec(description)) !== null) {
     result.push(description.substring(index, match.index));
-    result.push({ href: trackerUrl.replace('%ID%', match[1]), text: match[0] });
+    const text = match[0];
+    const url = trackerUrl.replace('%ID%', match[1]);
+    result.push(`[${text}](${url})`);
     index = reIssueId.lastIndex;
   }
   if (index < description.length) {
     result.push(description.substring(index, description.length));
   }
-  return result.length > 0 ? result : description;
+  return result.length > 0 ? result.join('') : description;
 };
 
 class Description extends React.Component {
   static propTypes = {
     children: PropTypes.oneOfType([
       PropTypes.string,
-      PropTypes.arrayOf(
-        PropTypes.oneOfType([
-          PropTypes.string,
-          PropTypes.shape({
-            href: PropTypes.string.isRequired,
-            text: PropTypes.string.isRequired,
-          }),
-        ])
-      ),
+      PropTypes.arrayOf(PropTypes.string),
     ]),
     options: PropTypes.shape({
       trackerUrl: PropTypes.string,
@@ -45,26 +42,11 @@ class Description extends React.Component {
     options: {},
   };
 
-  renderArray(entries) {
-    return entries.map(
-      (e, index) =>
-        typeof e === 'string'
-          ? <span key={index}>{e}</span>
-          : <a key={index} href={e.href} target="_blank">{e.text}</a>
-    );
-  }
-
   render() {
     const { children, options } = this.props;
     const { trackerUrl, pattern } = options;
     const description = getTrackerLinks(children, pattern, trackerUrl);
-    return (
-      <em>
-        {Array.isArray(description)
-          ? this.renderArray(description)
-          : description}
-      </em>
-    );
+    return <Markdown source={description} options={mdOptions} />;
   }
 }