+++ src/components/Panels.js
@@ -6,21 +6,13 @@ import Story from './Story';
 export class PropsPanel extends BasePanel {
   render() {
     const { data } = this.state;
-    return (
-      <div>
-        {data && <Doc data={data} />}
-      </div>
-    );
+    return <div>{data && <Doc data={data} />}</div>;
   }
 }
 
 export class StoryPanel extends BasePanel {
   render() {
     const { data } = this.state;
-    return (
-      <div>
-        {data && <Story data={data} />}
-      </div>
-    );
+    return <div>{data && <Story data={data} />}</div>;
   }
 }